new classes and tables for two markets
